modules/location_producer/requirements.txt
grpcio-tools==1.50.0
kafka-python==2.0.2